🚚 Rename islandProps.ts to enums.ts and move to root
