+++ heardle_server/enums.ts
