+++ heardle_server/islands/root.tsx
@@ -5,7 +5,7 @@ import GuessBar from "./guess-bar.tsx";
 import SongBar from "./song-bar.tsx";
 import ProgressBlock from "./progress-block.tsx";
 import { PastGuess } from "./islandProps.d.ts";
-import { guessResult } from "./islandProps.ts";
+import { guessResult } from "../enums.ts";
 import {checkStorageAvailable, hasWon} from "../helpers.tsx";
 import ShareButton from "./share-button.tsx";
