+++ heardle_server/islands/share-button.tsx
@@ -3,7 +3,7 @@ import { useState } from "preact/hooks";
 
 import Button from "../components/Button.tsx";
 import {PastGuess} from "./islandProps.d.ts";
-import {guessResult} from "./islandProps.ts";
+import {guessResult} from "../enums.ts";
 
 export default function ShareButton({ gameIsOver, gameTitle, currentDate, history } : { gameIsOver: boolean, gameTitle: string, currentDate: Date, history: Signal<PastGuess[]> }) {
   const [successfulCopy, setSuccessfulCopy] = useState(false);
